No longer navigate to configure page ever from kbar
I don't know what the actual problem is, I can't reproduce, but it seems
for some users kbar is always sending to the configure page, which just
isn't expected ever really. We'll now just take them to the run page as
if they clicked on it from the dashboard no matter what and let them
attempt to run it.